creating handlers and removig some path parameters
