+++ backend/backend/api/app.py
@@ -1,8 +1,8 @@
 from fastapi import FastAPI, HTTPException
 from fastapi.exception_handlers import http_exception_handler
 
-from backend.errors import room as room_errors
 from .routes import room, player
+from backend.api.handlers import room as room_errors
 
 app = FastAPI()
 app.include_router(room.router)
